+++ CHANGELOG.rst
@@ -3,6 +3,9 @@ Changelog
 
 ## 0.19.1
   * Apply black and isort
+  * Add ``--log-level`` option to the main CLI.
+  * Include app in coverage calculation.
+  * Add tests for app/sonata.
 
 ## 0.19.0
   * Enforce layer to be a string (NSETM-2261)